pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stage1_5"
version = "0.1.0"
description = "Stage 1.5 latent separability audit pipeline (Accent × Speaker)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "OpenCode", email = "ops@example.com"}
]
dependencies = [
  "numpy>=1.26",
  "scipy>=1.11",
  "pandas>=2.1",
  "scikit-learn>=1.4",
  "librosa>=0.10",
  "praat-parselmouth>=0.4",
  "soundfile>=0.12",
  "torch>=2.2",
  "torchaudio>=2.2",
  "speechbrain>=0.5.16",
  "s3prl>=0.4.12",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "tqdm>=4.66",
  "typer>=0.12",
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "matplotlib>=3.8",
  "seaborn>=0.13"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.1",
  "mypy>=1.9",
  "types-PyYAML",
  "types-tqdm"
]

[project.scripts]
stage1_5 = "stage1_5.cli:app"

[tool.setuptools.package-data]
stage1_5 = ["reporting/templates/*.md"]

[tool.pytest.ini_options]
addopts = "-ra"
pythonpath = ["."]
